refactor: formbuilder settings
